refactor: migrate from setup() to pyproject.toml
