+++ install/README.md
@@ -1,0 +1,3 @@
+# Install
+
+These are pip pre-build steps to collect all static assets that are part of the swprw ReactJS application that runs inside of this swpwrxblock. **Managed by pyproject.toml**
