+++ pyproject.toml
@@ -1,13 +1,16 @@
 [build-system]
 requires = ["setuptools", "wheel"]
-build-backend = "setuptools.build_meta"
+build-backend = "install.build_backend.CustomBuildBackend"
 
 [project]
 name = "stepwise-power-xblock"
-version = "0.1.0"  # Replace with the actual version
+version = "18.0.1"
 description = "Stepwise Power XBlock"
 license = { text = "MIT" }
 dependencies = ["XBlock", "requests"]
+
+[project.entry-points."xblock.v1"]
+swpwrxblock = "swpwrxblock:SWPWRXBlock"
 
 [tool.setuptools]
 packages = ["swpwrxblock"]
@@ -15,6 +18,3 @@ include-package-data = true
 
 [tool.setuptools.package-data]
 "swpwrxblock" = ["static/*", "public/*", "translations/*"]
-
-[tool.setuptools.entry-points."xblock.v1"]
-swpwrxblock = "swpwrxblock:SWPWRXBlock"
